Moved YTdlp class to Framework layer under the external package. Reformatted some code. Fixed character encoding issue in RadioInteractions.
