+++ src/main/java/abr/song_downloader_abr/SongDownloader.java
@@ -1,13 +1,14 @@
 package abr.song_downloader_abr;
 
+import abr.song_abr.SongDAOInput;
 import com.jayway.jsonpath.DocumentContext;
 import com.jayway.jsonpath.JsonPath;
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoClients;
 import ds.DatabaseInitializer;
-import abr.song_abr.SongDAOInput;
 import ds.song_ds.SongDAOInputImpl;
 import entities.Song;
+import framework.external.YTdlp;
 
 import java.io.IOException;
 import java.nio.file.FileVisitOption;
@@ -17,7 +18,16 @@ import java.util.Scanner;
 
 public class SongDownloader {
 
-    public static void main(String[] args) throws Exception {
+    public static Song readJSON(DocumentContext dc) throws IOException {
+        String id = dc.read("$['id']");
+        String name = dc.read("$['title']");
+        int duration = dc.read("$['duration']");
+
+        return new Song(name, id, duration);
+    }
+
+//    TODO: rename from main when finished project
+    public void main(String[] args) throws Exception {
 //        Single video or playlist link
 //        https://www.youtube.com/watch?v=gWo12TtN9Kk
         System.out.println("Enter youtube link: ");
@@ -29,7 +39,7 @@ public class SongDownloader {
         moveToDatabase();
     }
 
-    public static void moveToDatabase() throws Exception {
+    public void moveToDatabase() throws Exception {
 //        download json and mp3 to directory using yt-dlp
 //        iterate over json files in directory
 //        create Song object if json file does not belong to playlist
@@ -38,14 +48,11 @@ public class SongDownloader {
         String uri = "mongodb://root:rootpassword@localhost:27017";
         DatabaseInitializer.init();
 
-//        TODO: change path for laptop
         try (MongoClient mongoClient = MongoClients.create(uri)) {
             SongDAOInput songDAOin = new SongDAOInputImpl(mongoClient);
 
             Files.walk(Paths.get("C:\\Users\\allen\\Desktop\\csc207\\course-project-shazamify\\build\\songs"), FileVisitOption.FOLLOW_LINKS).filter(t ->
-            {
-                return t.toString().endsWith(".info.json");
-            }).forEach(path -> {
+                    t.toString().endsWith(".info.json")).forEach(path -> {
                 try {
                     String content = Files.readString(path);
                     DocumentContext jsonContext = JsonPath.parse(content);
@@ -60,13 +67,5 @@ public class SongDownloader {
 
             });
         }
-    }
-
-    public static Song readJSON(DocumentContext dc) throws IOException {
-        String id = dc.read("$['id']");
-        String name = dc.read("$['title']");
-        int duration = dc.read("$['duration']");
-
-        return new Song(name, id, duration);
     }
 }
