+++ src/main/java/interface_adaptors/SearchResultsViewModel.java
@@ -1,12 +1,12 @@
 package interface_adaptors;
 
 import entities.Song;
+import entities.user_entities.User;
 import framework.buttons.ButtonSearchAlbums;
 import framework.buttons.ButtonSearchSongs;
 import framework.buttons.ButtonSearchUsers;
 import framework.items.SearchSongItem;
 import framework.items.SearchUserItem;
-import user.entities.User;
 
 import javax.swing.*;
 import java.awt.*;
